Moved the new functions to utils.go to simplify changes.
